% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_check_code}
\alias{cb_check_code}
\title{Check code if correct}
\usage{
cb_check_code(code_check, code_secret)
}
\arguments{
\item{code_check}{Code to check}

\item{code_secret}{Secret code}
}
\value{
list
}
\description{
Check code if correct
}
