% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_clean_code}
\alias{cb_clean_code}
\title{Clean code}
\usage{
cb_clean_code(code, code_length = 4)
}
\arguments{
\item{code}{Code}

\item{code_length}{Length of code}
}
\value{
clean code
}
\description{
Clean code
}
