% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_input_code}
\alias{cb_input_code}
\title{Input code}
\usage{
cb_input_code(step = 1, code_length = 4, color_list = c("R", "B"))
}
\arguments{
\item{step}{Step}

\item{code_length}{Length of code}

\item{color_list}{List of available colors}
}
\value{
code
}
\description{
Input code
}
