% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_intro}
\alias{cb_intro}
\title{Show Code Breaker Intro}
\usage{
cb_intro(name = NULL)
}
\arguments{
\item{name}{Player name}
}
\value{
Prints sprite in console
}
\description{
Show Code Breaker Intro
}
