% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_play_game}
\alias{cb_play_game}
\title{Play a code breaker game}
\usage{
cb_play_game(colors = NA, sound = TRUE, name = NULL)
}
\arguments{
\item{colors}{Colors that can be used in game}

\item{sound}{Play sounds?}

\item{name}{Player name}
}
\value{
list
}
\description{
Play a code breaker game
}
