% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_race_cup}
\alias{cb_race_cup}
\title{Show Code Breaker Race Cup}
\usage{
cb_race_cup(name = NULL)
}
\arguments{
\item{name}{Player name}
}
\value{
Prints sprite in console
}
\description{
Show Code Breaker Race Cup
}
