% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_select_colors}
\alias{cb_select_colors}
\title{Selct Colors}
\usage{
cb_select_colors(colors = NA, name = NULL)
}
\arguments{
\item{colors}{Selected colors}

\item{name}{Player name}
}
\value{
Print selected colors in console
}
\description{
Selct Colors
}
