% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_show_color}
\alias{cb_show_color}
\title{Show color in console}
\usage{
cb_show_color(color)
}
\arguments{
\item{color}{Color}
}
\value{
Prints color in console
}
\description{
Show color in console
}
