% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{cb_success}
\alias{cb_success}
\title{Show Code Breaker Success}
\usage{
cb_success(name = NULL)
}
\arguments{
\item{name}{Player name}
}
\value{
Prints sprite in console
}
\description{
Show Code Breaker Success
}
