% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebreaker.R
\name{codebreaker}
\alias{codebreaker}
\title{Play codebreaker game}
\usage{
codebreaker(sound = TRUE, name = NULL)
}
\arguments{
\item{sound}{Play sounds?}

\item{name}{Name of player}
}
\value{
list
}
\description{
Play codebreaker game
}
\examples{
## Start game (in interactive R sessions)
if (interactive())  {
   codebreaker()
}
}
